example migration of single unit test
dropping rewire ftw
