+++ .eslintrc.js
@@ -1,6 +1,9 @@
 module.exports = {
   parser: 'babel-eslint',
   extends: 'standard',
+  env: {
+    jest: true
+  },
   plugins: [
     'standard',
     'promise'
